Revert "Fix bare URLs are getting cleared on Markdown"
This reverts commit 7fe40f8f95672d50f147f24234b9e0095c003cc8.
