+++ richeditor-compose/src/commonTest/kotlin/com.mohamedrejeb.richeditor/model/RichTextStateTest.kt
@@ -229,7 +229,6 @@ class RichTextStateTest {
         assertEquals(richTextState.currentSpanStyle, SpanStyle(fontWeight = FontWeight.Bold))
     }
 
-    @OptIn(ExperimentalRichTextApi::class)
     @Test
     fun testClearSpanStyles() {
         val richTextState = RichTextState(
@@ -422,7 +421,6 @@ class RichTextStateTest {
         assertEquals(defaultSpan, richTextState.currentRichSpanStyle)
     }
 
-    @OptIn(ExperimentalRichTextApi::class)
     @Test
     fun testGetSpanStyle() {
         val richTextState = RichTextState(
