+++ richeditor-compose/src/commonTest/kotlin/com.mohamedrejeb.richeditor/parser/markdown/RichTextStateMarkdownParserTest.kt
@@ -13,7 +13,180 @@ import com.mohamedrejeb.richeditor.paragraph.RichParagraph
 import kotlin.test.Test
 import kotlin.test.assertEquals
 
-class RichTextStateMarkdownParserDecodeTest {
+class RichTextStateMarkdownParserTest {
+
+    /**
+     * Encode tests
+     */
+
+    @Test
+    fun testBold() {
+        val markdown = "**Hello World!**"
+        val expectedText = "Hello World!"
+        val state = RichTextStateMarkdownParser.encode(markdown)
+        val actualText = state.annotatedString.text
+
+        assertEquals(
+            expected = expectedText,
+            actual = actualText,
+        )
+
+        assertEquals(
+            expected = SpanStyle(fontWeight = FontWeight.Bold),
+            actual = state.richParagraphList.first().children.first().spanStyle
+        )
+    }
+
+    @Test
+    fun testBoldWithNestedItalic() {
+        val markdown = "**Hello *World!***"
+        val expectedText = "Hello World!"
+        val state = RichTextStateMarkdownParser.encode(markdown)
+        val actualText = state.annotatedString.text
+
+        assertEquals(
+            expected = expectedText,
+            actual = actualText,
+        )
+
+        val firstChild = state.richParagraphList.first().children.first()
+        val secondChild = firstChild.children.first()
+
+        assertEquals(
+            expected = SpanStyle(fontWeight = FontWeight.Bold),
+            actual = firstChild.spanStyle
+        )
+
+        assertEquals(
+            expected = SpanStyle(fontStyle = FontStyle.Italic),
+            actual = secondChild.spanStyle
+        )
+    }
+
+    @Test
+    fun testUnderline() {
+        val markdown = "<u>Hello World!</u>"
+        val expectedText = "Hello World!"
+        val state = RichTextStateMarkdownParser.encode(markdown)
+        val actualText = state.annotatedString.text
+
+        assertEquals(
+            expected = expectedText,
+            actual = actualText,
+        )
+
+        assertEquals(
+            expected = SpanStyle(textDecoration = TextDecoration.Underline),
+            actual = state.richParagraphList.first().children.first().spanStyle
+        )
+    }
+
+    @Test
+    fun testLineThrough() {
+        val markdown = "~~Hello World!~~"
+        val expectedText = "Hello World!"
+        val state = RichTextStateMarkdownParser.encode(markdown)
+        val actualText = state.annotatedString.text
+
+        assertEquals(
+            expected = expectedText,
+            actual = actualText,
+        )
+
+        assertEquals(
+            expected = SpanStyle(textDecoration = TextDecoration.LineThrough),
+            actual = state.richParagraphList.first().children.first().spanStyle
+        )
+    }
+
+    @Test
+    fun testEncodeLineBreak() {
+        val markdown = """
+            Hello
+            <br>
+            
+            
+            
+            World!
+        """.trimIndent()
+
+        val state = RichTextStateMarkdownParser.encode(markdown)
+
+        assertEquals(
+            expected = 4,
+            actual = state.richParagraphList.size,
+        )
+
+        state.setMarkdown(
+            """
+            Hello
+            
+            
+            
+            World!
+        """.trimIndent()
+        )
+
+        assertEquals(
+            expected = 3,
+            actual = state.richParagraphList.size,
+        )
+
+        state.setMarkdown(
+            """
+            Hello
+            
+            <br>
+            <br>
+            
+            World!
+        """.trimIndent()
+        )
+
+        assertEquals(
+            expected = 5,
+            actual = state.richParagraphList.size,
+        )
+    }
+
+    @Test
+    fun testEncodeMarkdownWithSingleDollar() {
+        val markdown = "Hello World $100!"
+        val expectedText = "Hello World $100!"
+        val state = RichTextStateMarkdownParser.encode(markdown)
+        val actualText = state.annotatedString.text
+
+        assertEquals(
+            expected = expectedText,
+            actual = actualText,
+        )
+    }
+
+    @Test
+    fun testEncodeMarkdownWithDoubleDollar() {
+        val markdown = "Hello World $$100!"
+        val expectedText = "Hello World $$100!"
+        val state = RichTextStateMarkdownParser.encode(markdown)
+        val actualText = state.annotatedString.text
+
+        assertEquals(
+            expected = expectedText,
+            actual = actualText,
+        )
+    }
+
+    @Test
+    fun testEncodeMarkdownWithInlineMath() {
+        val markdown = "Hello World \$100$!"
+        val expectedText = "Hello World 100!"
+        val state = RichTextStateMarkdownParser.encode(markdown)
+        val actualText = state.annotatedString.text
+
+        assertEquals(
+            expected = expectedText,
+            actual = actualText,
+        )
+    }
 
     /**
      * Decode tests
